data/backup/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2

# Data processing
pandas==2.2.2
regex==2025.7.34

# Database
sqlalchemy==2.0.43
psycopg2-binary>=2.9.9
cloud-sql-python-connector[pg8000]

# Authentication & Security
authlib==1.3.1       
PyJWT==2.8.0

# Web server & utilities
gunicorn>=20.1.0
python-dotenv==1.1.1
jinja2==3.1.2
python-multipart==0.0.20
httpx==0.26.0

# External integrations
gspread==6.2.1  # Google Sheets integration - used in utils.py
inscriptis==2.6.0
pymupdf==1.26.4
groq==0.31.0

itsdangerous==2.2.0
gcsfs==2024.10.0
psutil==6.1.0
# Removed potentially unused packages:
# protobuf==3.20.1 - Protocol Buffers (not found in codebase)
# Werkzeug==2.2.2 - WSGI utilities (not found in codebase)
# itsdangerous==2.2.0 - Digital signatures (not found in codebase)
# python-jose[cryptography]==3.3.0 - JWT library (duplicate of PyJWT)
# urllib3<2.0 - HTTP library (httpx provides this functionality)